Addresses #1 : Autogenerate ID numbers for objects with 'id: null' in their YAML
This also changes the way scripts are serialized. Scripts are serialized with a magic number, but I couldn't figure out how it was calculated. Thankfully, the editor does not appear to care about the number - at least, VX Ace doesn't - so I just got rid of it, and used the same sequential ID numbering that is used in the rest of the YAML structures. Works great.

Items with an ID already set should be preserved, so you should be safe to do this without worrying about references to item ID numbers changing in events & etc. In other words, this should be collaboration-safe.